move properties to dedicated directory
