+++ javascript/experiments/use-webpack-and-swc/src/properties/getSize.js
